rust meetup work

+++ tuning_systems/src/tuning_systems/tone.rs
@@ -15,12 +15,12 @@ pub struct Tone {
 
 impl Tone {
     // #[cfg_attr(feature = "wasm-bindgen", wasm_bindgen(constructor))]
-    pub fn new(tuning_system: TuningSystem, tone_index: usize) -> Tone {
+    pub fn new(tuning_system: TuningSystem, tone_index: u32) -> Tone {
         Tone::new_with_octave_size(tuning_system, tone_index)
     }
 
     // #[cfg_attr(feature = "wasm-bindgen", wasm_bindgen(constructor))]
-    pub(crate) fn new_with_octave_size(tuning_system: TuningSystem, tone_index: usize) -> Tone {
+    pub(crate) fn new_with_octave_size(tuning_system: TuningSystem, tone_index: u32) -> Tone {
         let name = tuning_system.get_tone_name(tone_index);
         let fraction = tuning_system.get_fraction(tone_index);
 
@@ -37,7 +37,7 @@ impl Tone {
     }
 
     pub fn cents(&self) -> f64 {
-        let reference_freq: f64 = equal_temperament(self.tone_index, self.tuning_system.size()).into();
+        let reference_freq: f64 = equal_temperament(self.tone_index as u32, self.tuning_system.size() as u32).into();
         let comparison_freq: f64 = self.frequency();
         1200f64 * (comparison_freq / reference_freq).log2()
     }
